+++ app/src/main/java/com/alphawallet/app/repository/TokenRepository.java
@@ -98,7 +98,7 @@ public class TokenRepository implements TokenRepositoryType {
     private static final int NODE_COMMS_ERROR = -1;
     private static final int CONTRACT_BALANCE_NULL = -2;
 
-    private final Map<Integer, Web3j> web3jNodeServers;
+    private final Map<Long, Web3j> web3jNodeServers;
     private AWEnsResolver ensResolver;
 
     public TokenRepository(
@@ -124,7 +124,7 @@ public class TokenRepository implements TokenRepositoryType {
         web3jNodeServers.put(networkInfo.chainId, Web3j.build(publicNodeService));
     }
 
-    private Web3j getService(int chainId)
+    private Web3j getService(long chainId)
     {
         if (!web3jNodeServers.containsKey(chainId))
         {
@@ -194,14 +194,14 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public TokenCardMeta[] fetchTokenMetasForUpdate(Wallet wallet, List<Integer> networkFilters)
+    public TokenCardMeta[] fetchTokenMetasForUpdate(Wallet wallet, List<Long> networkFilters)
     {
         if (networkFilters == null) networkFilters = Collections.emptyList(); //if filter null, return all networks
         return localSource.fetchTokenMetasForUpdate(wallet, networkFilters);
     }
 
     @Override
-    public Single<TokenCardMeta[]> fetchTokenMetas(Wallet wallet, List<Integer> networkFilters,
+    public Single<TokenCardMeta[]> fetchTokenMetas(Wallet wallet, List<Long> networkFilters,
                                                    AssetDefinitionService svs)
     {
         if (networkFilters == null) networkFilters = Collections.emptyList(); //if filter null, return all networks
@@ -210,21 +210,21 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Single<TokenCardMeta[]> fetchAllTokenMetas(Wallet wallet, List<Integer> networkFilters, String searchTerm) {
+    public Single<TokenCardMeta[]> fetchAllTokenMetas(Wallet wallet, List<Long> networkFilters, String searchTerm) {
         if (networkFilters == null) networkFilters = Collections.emptyList(); //if filter null, return all networks
         return localSource
                 .fetchAllTokenMetas(wallet, networkFilters, searchTerm);
     }
 
     @Override
-    public Single<Token[]> fetchTokensThatMayNeedUpdating(String walletAddress, List<Integer> networkFilters) {
+    public Single<Token[]> fetchTokensThatMayNeedUpdating(String walletAddress, List<Long> networkFilters) {
         if (networkFilters == null) networkFilters = Collections.emptyList(); //if filter null, return all networks
         return localSource
                 .fetchAllTokensWithNameIssue(walletAddress, networkFilters);
     }
 
     @Override
-    public Single<ContractAddress[]> fetchAllTokensWithBlankName(String walletAddress, List<Integer> networkFilters) {
+    public Single<ContractAddress[]> fetchAllTokensWithBlankName(String walletAddress, List<Long> networkFilters) {
         if (networkFilters == null) networkFilters = Collections.emptyList(); //if filter null, return all networks
         return localSource
                 .fetchAllTokensWithBlankName(walletAddress, networkFilters);
@@ -243,7 +243,7 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Single<BigInteger> fetchLatestBlockNumber(int chainId)
+    public Single<BigInteger> fetchLatestBlockNumber(long chainId)
     {
         return Single.fromCallable(() -> {
             try
@@ -260,7 +260,7 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Token fetchToken(int chainId, String walletAddress, String address)
+    public Token fetchToken(long chainId, String walletAddress, String address)
     {
         Wallet wallet = new Wallet(walletAddress);
         return localSource.fetchToken(chainId, wallet, address);
@@ -297,7 +297,7 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Single<BigDecimal> fetchChainBalance(String walletAddress, int chainId)
+    public Single<BigDecimal> fetchChainBalance(String walletAddress, long chainId)
     {
         Token baseToken = fetchToken(chainId, walletAddress, walletAddress);
         return updateTokenBalance(walletAddress, baseToken);
@@ -365,7 +365,7 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Single<String> resolveENS(int chainId, String ensName)
+    public Single<String> resolveENS(long chainId, String ensName)
     {
         if (ensResolver == null) ensResolver = new AWEnsResolver(TokenRepository.getWeb3jService(MAINNET_ID), context);
         return ensResolver.resolveENSAddress(ensName);
@@ -384,13 +384,13 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Single<TokenInfo> update(String contractAddr, int chainId)
+    public Single<TokenInfo> update(String contractAddr, long chainId)
     {
         return setupTokensFromLocal(contractAddr, chainId);
     }
 
     @Override
-    public String getTokenImageUrl(int networkId, String address)
+    public String getTokenImageUrl(long networkId, String address)
     {
         return localSource.getTokenImageUrl(networkId, address);
     }
@@ -468,7 +468,7 @@ public class TokenRepository implements TokenRepositoryType {
         });
     }
 
-    private BigDecimal checkUint256Balance(@NonNull Wallet wallet, int chainId, String tokenAddress)
+    private BigDecimal checkUint256Balance(@NonNull Wallet wallet, long chainId, String tokenAddress)
     {
         BigDecimal balance = BigDecimal.valueOf(-1);
 
@@ -557,12 +557,12 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Single<TokenTicker> getEthTicker(int chainId)
+    public Single<TokenTicker> getEthTicker(long chainId)
     {
         return Single.fromCallable(() -> tickerService.getEthTicker(chainId));
     }
 
-    private BigDecimal getEthBalance(Wallet wallet, int chainId)
+    private BigDecimal getEthBalance(Wallet wallet, long chainId)
     {
         //in case chain has an override
         if (EthereumNetworkRepository.getChainOverrideAddress(chainId).length() > 0)
@@ -586,7 +586,7 @@ public class TokenRepository implements TokenRepositoryType {
         }
     }
 
-    private List<BigInteger> getBalanceArray875(Wallet wallet, int chainId, String tokenAddress) {
+    private List<BigInteger> getBalanceArray875(Wallet wallet, long chainId, String tokenAddress) {
         List<BigInteger> result = new ArrayList<>();
         result.add(BigInteger.valueOf(NODE_COMMS_ERROR));
         try
@@ -638,7 +638,7 @@ public class TokenRepository implements TokenRepositoryType {
         return result;
     }
 
-    private List<BigInteger> getBalanceArray721Ticket(Wallet wallet, int chainId, String tokenAddress) {
+    private List<BigInteger> getBalanceArray721Ticket(Wallet wallet, long chainId, String tokenAddress) {
         List<BigInteger> result = new ArrayList<>();
         result.add(BigInteger.valueOf(NODE_COMMS_ERROR));
         try
@@ -991,7 +991,7 @@ public class TokenRepository implements TokenRepositoryType {
      * @return
      */
     private String callCustomNetSmartContractFunction(
-            Function function, String contractAddress, Wallet wallet, int chainId)  {
+            Function function, String contractAddress, Wallet wallet, long chainId)  {
         String encodedFunction = FunctionEncoder.encode(function);
 
         try
@@ -1073,7 +1073,7 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public Single<ContractLocator> getTokenResponse(String address, int chainId, String method)
+    public Single<ContractLocator> getTokenResponse(String address, long chainId, String method)
     {
         return Single.fromCallable(() -> {
             ContractLocator contractLocator = new ContractLocator(INVALID_CONTRACT, chainId);
@@ -1098,7 +1098,7 @@ public class TokenRepository implements TokenRepositoryType {
         });
     }
 
-    private Single<TokenInfo> setupTokensFromLocal(String address, int chainId) //pass exception up the chain
+    private Single<TokenInfo> setupTokensFromLocal(String address, long chainId) //pass exception up the chain
     {
         return Single.fromCallable(() -> {
             NetworkInfo network = ethereumNetworkRepository.getNetworkByChain(chainId);
@@ -1268,12 +1268,12 @@ public class TokenRepository implements TokenRepositoryType {
     }
 
     @Override
-    public void addImageUrl(int networkId, String address, String imageUrl)
+    public void addImageUrl(long networkId, String address, String imageUrl)
     {
         localSource.storeTokenUrl(networkId, address, imageUrl);
     }
 
-    public static Web3j getWeb3jService(int chainId)
+    public static Web3j getWeb3jService(long chainId)
     {
         OkHttpClient okClient = new OkHttpClient.Builder()
                 .connectTimeout(10, TimeUnit.SECONDS)
@@ -1286,7 +1286,7 @@ public class TokenRepository implements TokenRepositoryType {
         return Web3j.build(publicNodeService);
     }
 
-    public static String callSmartContractFunction(int chainId,
+    public static String callSmartContractFunction(long chainId,
                                   Function function, String contractAddress, String walletAddr)
     {
         String encodedFunction = FunctionEncoder.encode(function);
@@ -1312,7 +1312,7 @@ public class TokenRepository implements TokenRepositoryType {
         return null;
     }
 
-    public static List callSmartContractFunctionArray(int chainId,
+    public static List callSmartContractFunctionArray(long chainId,
                                 Function function, String contractAddress, String walletAddr)
     {
         try
