+++ app/src/main/java/com/alphawallet/app/repository/entity/RealmTransaction.java
@@ -17,7 +17,7 @@ public class RealmTransaction extends RealmObject {
     private String gasUsed;
     private String input;
     private String error;
-    private int chainId;
+    private long chainId;
     private long expectedCompletion;
     private String contractAddress; // this is so we can efficiently lookup transactions relating to contracts,
                                     // if we discovered them using the Etherscan 'Transfers' API.
@@ -121,12 +121,12 @@ public class RealmTransaction extends RealmObject {
         this.error = error;
     }
 
-    public int getChainId()
+    public long getChainId()
     {
         return chainId;
     }
 
-    public void setChainId(int chainId)
+    public void setChainId(long chainId)
     {
         this.chainId = chainId;
     }
