+++ app/src/main/java/com/alphawallet/app/ui/Erc20DetailActivity.java
@@ -88,7 +88,7 @@ public class Erc20DetailActivity extends BaseActivity implements StandardFunctio
         {
             symbol = savedInstanceState.getString(C.EXTRA_ACTION_NAME);
             wallet = savedInstanceState.getParcelable(WALLET);
-            int chainId = savedInstanceState.getInt(C.EXTRA_CHAIN_ID, MAINNET_ID);
+            long chainId = savedInstanceState.getLong(C.EXTRA_CHAIN_ID, MAINNET_ID);
             token = viewModel.getTokensService().getToken(chainId, savedInstanceState.getString(C.EXTRA_ADDRESS));
         }
         else
@@ -173,7 +173,7 @@ public class Erc20DetailActivity extends BaseActivity implements StandardFunctio
         symbol = getIntent().getStringExtra(C.EXTRA_SYMBOL);
         symbol = symbol == null ? ETH_SYMBOL : symbol;
         wallet = getIntent().getParcelableExtra(WALLET);
-        int chainId = getIntent().getIntExtra(C.EXTRA_CHAIN_ID, MAINNET_ID);
+        long chainId = getIntent().getLongExtra(C.EXTRA_CHAIN_ID, MAINNET_ID);
         token = viewModel.getTokensService().getTokenOrBase(chainId, getIntent().getStringExtra(C.EXTRA_ADDRESS));
         tokenMeta = new TokenCardMeta(token);
         viewModel.checkForNewScript(token);
@@ -185,7 +185,7 @@ public class Erc20DetailActivity extends BaseActivity implements StandardFunctio
         super.onSaveInstanceState(outState);
         outState.putString(C.EXTRA_SYMBOL, symbol);
         outState.putParcelable(WALLET, wallet);
-        outState.putInt(C.EXTRA_CHAIN_ID, token.tokenInfo.chainId);
+        outState.putLong(C.EXTRA_CHAIN_ID, token.tokenInfo.chainId);
         outState.putString(C.EXTRA_ADDRESS, token.getAddress());
     }
