+++ app/src/main/java/com/alphawallet/app/ui/SendActivity.java
@@ -110,7 +110,7 @@ public class SendActivity extends BaseActivity implements AmountReadyCallback, S
                 .get(SendViewModel.class);
 
         String contractAddress = getIntent().getStringExtra(C.EXTRA_CONTRACT_ADDRESS);
-        int currentChain = getIntent().getIntExtra(C.EXTRA_NETWORKID, MAINNET_ID);
+        long currentChain = getIntent().getLongExtra(C.EXTRA_NETWORKID, MAINNET_ID);
         wallet = getIntent().getParcelableExtra(WALLET);
         token = viewModel.getToken(currentChain, getIntent().getStringExtra(C.EXTRA_ADDRESS));
         QRResult result = getIntent().getParcelableExtra(C.EXTRA_AMOUNT);
@@ -147,7 +147,7 @@ public class SendActivity extends BaseActivity implements AmountReadyCallback, S
         }
     }
 
-    private boolean checkTokenValidity(int currentChain, String contractAddress)
+    private boolean checkTokenValidity(long currentChain, String contractAddress)
     {
         if (token == null || token.tokenInfo == null)
         {
@@ -424,7 +424,7 @@ public class SendActivity extends BaseActivity implements AmountReadyCallback, S
         }
     }
 
-    private void showChainChangeDialog(int chainId)
+    private void showChainChangeDialog(long chainId)
     {
         if (dialog != null && dialog.isShowing()) dialog.dismiss();
         dialog = new AWalletAlertDialog(this);
